+++ src/helpers/cli/generateSerializerContent.ts
@@ -88,7 +88,7 @@ export default function generateSerializerContent({
       if (name === undefined) return ''
       if (['belongs_to', 'has_one', 'has_many'].includes(type as any)) return ''
 
-      return `\n    ${attribute(name, type, attr, stiBaseSerializer)}`
+      return `\n    ${attribute(modelClassName, name, type, attr, stiBaseSerializer)}`
     })
     .join('')}`
 
@@ -107,9 +107,15 @@ ${defaultSerializer.replace(serializerClassName, adminSerializerClassName).repla
 `
 }
 
-function attribute(name: string, type: string | undefined, attr: string, stiBaseSerializer: boolean) {
+function attribute(
+  modelClassName: string,
+  name: string,
+  type: string | undefined,
+  attr: string,
+  stiBaseSerializer: boolean
+) {
   if (name === 'type' && stiBaseSerializer) {
-    return `.attribute('type', { openapi: { type: 'string', enum: [StiChildClass.sanitizedName] } })`
+    return `.attribute('type', { openapi: { type: 'string', enum: [(StiChildClass ?? ${modelClassName}).sanitizedName] } })`
   }
 
   switch (type) {